Data.BStream

Documentation

newtype BStream a

Infinite complete binary trees

Constructors

BStream

Fields

runBStream :: Cofree Pair a

bools :: Pair (BStream Bool)

paths bools is the list of all infinite binary sequences. Well, kind of.. it'a really just repeat False, but if we could complete collecting infinite series in finite time, we could list them all.

scanMap :: Semigroup s => (a -> s) -> BStream a -> BStream s

Scan along the results of a BStream with a Semigroup and a method to convert to it.

stepMap :: (a -> b -> b) -> (a -> b) -> BStream a -> BStream b

Use a map and step to scan a BStream

blevels :: BStream a -> [[a]]

All levels of a BStream (infinite)

blevelN :: Int -> BStream a -> [a]

The n-th level of a BStream