Data.Cycle.Find

Documentation

foldl1Maybe :: Foldable t => (b -> a -> b) -> (a -> b) -> t a -> Maybe b

Get the initial state from the first element of a Foldable or return Nothing

goal :: ()

Goal:

 Goal:
   Given a list, function, series, etc.
   1. If there exists xs /= input | and (zipWith (==) (cycle xs) input)
      Output the shortest, else Nothing.
   2. If there exists xs | exists (0 <= n) | and (zipWith (==) (cycle xs) (drop n input) && n + 2 * length xs <= length input
      (i.e. it eventually cycles, and the cycle repeats fully at least once)
      Output the shortest, else Nothing.

 - If we end at anything other than a single value repeated, we recurse.

 So:
   Single value repeated n times
   Values repeated some number of times, recursed on their cycles

 Ahh!
   I was thinking we didn't need to store the entire previous list, we don't!
     We should just store a seq for the current cycle, since fast cons and fast single-element rotation.

 When a cycle ends, if there were no previous non-equal values, then we'll need to recurse.

 - If we keep a current-label state and add a label to each value, all value sets after the first will be a range [0..n]
   * Since the value sets are monotonically decreasing in size, we can use a single allocation to represent all of the maps

data Cycle a

A representation of a finite cycle (the list should never be infinite)

Constructors

Cycle

Fields

cycleLen :: Int

cycleSeries :: [a]

data Cyclic a

Constructors

Cyclic

Fields

cyclicLen :: Int

cyclicCycle :: Cycle a

cyclic :: Ord a => [a] -> Maybe [a]

Just the cycle that makes up the input or Nothing

 cyclic (take _ $ cycle xs) = Just xs
 cyclic _ = Nothing

boolMaybe :: (a -> Bool) -> a -> Maybe a

Just if the predicate returns True

isCycleOf :: Eq a => [a] -> [a] -> Bool

Is the second list a cycle of the first?

twoCycles :: [a] -> [(Bool, a)]

(have we reached two cycles?, value)

indexDiffs :: Ord a => [a] -> Map a (Int, [Int])

Indexed differences between locations of an element

intIndexDiffs :: [Int] -> IntMap (Indexed [Int])

indexDiffs specialized to Int

indices :: Eq a => a -> [a] -> [Int]

The indices of a value in a list

data Indexed a

An Int-indexed value

Constructors

Indexed

Fields

indexOf :: !Int

indexed :: a

uncurryi :: (Int -> a -> b) -> Indexed a -> b

uncurry for Indexed

curryi :: (Indexed a -> b) -> Int -> a -> b

curry for Indexed