Data.Semigroup.Free

Documentation

sizeNotes :: ()

Size notes:

 So a single element weighs ~400 bytes.
   Supposing a MArray has a 1-byte overhead per element,
   could we get a ~2 byte overhead per element if we stored every 400 elements in a MArray?
   Ehh.. I don't know if the price of copying arrays will offset the cost..

 What if we only access one cycle element at a time?
   In that case, it's probably better to simply carry around a list and cons onto it.

 What if we have a predictable cycle-size?
   In that case, it's probably best to use pre-allocated memory
   Possibly a pool of smaller pre-allocated chunks

notes :: ()

Note:

 On one hand, it seems nice to preserve the commutativity of the GappedIxCycle's
 On the other hand, it takes O(min n m) time to zip two GappedIxCycle's together.
 On the other hand, I'm not sure it needs to be anything other than a []SeqSet of period-1 cycles whose LCM is the given period. (that would give O(1) (<>))

class Applicative f => Semialternative f where

 GappedCycle a == Free Pair (Int, a)

data IntGappedCycle where IGCPure :: !Int -> !Int -> IntGappedCycle -- Period, Value IGCMappend :: !IntGappedCycle -> !IntGappedCycle -> IntGappedCycle

Alternative without empty

Minimal complete definition

(<!>)

Methods

(<!>) :: f a -> f a -> f a

saconcat :: NonEmpty (f a) -> f a

saconcatMap :: (a -> f b) -> NonEmpty a -> f b

newtype WrapSemialternative f a

Wrap a Semialternative to provide a Semigroup instance

Constructors

WrapSemialternative

Fields

runWrapSemialternative :: f a

class Semialternative f => SemigroupFree f where

A free Semigroup is a Semialternative that we can foldMap1 over (foldMap specialized to non-empty inputs) and convert to any other SemigroupFree (they're all isomorphic to each other).

Minimal complete definition

foldMap1 | semigroupFree

Methods

foldMap1 :: Semigroup s => (a -> s) -> f a -> s

semigroupFree :: SemigroupFree g => f a -> g a

data FreeSemigroup a where

From the triangle benchmakrs, this is the fastest. How fast? ~100x faster than NonEmpty, taking ~180 ms to construct and destruct a triangle with around a million elements.

For a little comparison, the average human blink of an eye is around 100-400 ms (according to WolframAlpha). I can process a trangular collection of a million elements in less then half the average time that it takes to blink an eye!

That's ~180 ns per element..

My computer has a 3.1 GHz processor. Assuming total utilization of a single core, we get a cycle-duration of ~0.323 ns. That gives us an estimate of 557 cycles per element.

That's not the greatest. I'll have to check CPU utilization for that benchmark to see if it's skewing my quick math. (267.2%?, not sure what to make of that. need to run without criterion to remove overhead) Though it's still not bad for a pure, polymorphic data type.

All in all, this data type is plenty fast for its primary purpose: quickly accumulating cycle-pieces when calculating the final cycle.

Constructors

PureFreeSemigroup :: a -> FreeSemigroup a

AppendFreeSemigroup :: FreeSemigroup a -> FreeSemigroup a -> FreeSemigroup a

toFreeSemigroup :: SemigroupFree f => f a -> FreeSemigroup a

semigroupFree specialized to FreeSemigroup

data FreeSemigroup' a where

It looks like this sucks around the [2^12..2^16] elements range

Constructors

PureFreeSemigroup' :: !a -> FreeSemigroup' a

AppendFreeSemigroup' :: !(FreeSemigroup' a) -> !(FreeSemigroup' a) -> FreeSemigroup' a

newtype FreePS a

Free-based FreeSemigroup

Constructors

FreePS

Fields

runFreePS :: Free Pair a

newtype FPS a

F-based FreeSemigroup

Constructors

FPS

Fields

runFPS :: F Pair a

newtype FSemigroup a

Partially applied foldMap1 wrapped in a newtype.

Constructors

FSemigroup

Fields

runFSemigroup :: forall s. Semigroup s => (a -> s) -> s

newtype FS a

FSemigroup with explicitly provided (<>).

Slightly faster in triangle benchmarks.

Constructors

FS

Fields

runFS :: forall s. (s -> s -> s) -> (a -> s) -> s

liftS2 :: (a -> b -> c) -> a -> b -> c

Lift function application to strict function application

newtype FS' a

Strict FSemigroup with explicitly provided (<>).

Constructors

FS'

Fields

runFS' :: forall s. (s -> s -> s) -> (a -> s) -> s

makeTriangle :: Semialternative f => Int -> f Int

A concatenated "triangle" of values:

 saconcat [1] <> saconcat [1, 2] <> saconcat [1, 2, 3] <> .. <> saconcat [1..n]

destructTriangle :: SemigroupFree f => Proxy f -> Int -> Int

Conveniently construct and Sum a triangle of SemigroupFree values for benchmarking.

makeLine :: Semialternative f => Int -> f Int

A concatenated line of values

 saconcat [1..n]

destructLine :: SemigroupFree f => Proxy f -> Int -> Int

Conveniently construct and Sum a line of SemigroupFree values for benchmarking.

proxyApInt :: proxy g -> Proxy (g Int)

Apply Int to a type inside a proxy

benchTriangle :: SemigroupFree f => Proxy f -> Int -> Benchmarkable

Benchmark destructTriangle

benchLine :: SemigroupFree f => Proxy f -> Int -> Benchmarkable

Benchmark destructLine

weighTriangle :: SemigroupFree f => String -> Proxy f -> Int -> Weigh ()

Weigh destructTriangle

weighLine :: SemigroupFree f => String -> Proxy f -> Int -> Weigh ()

Weigh destructLine

weighPow2s :: SemigroupFree f => String -> Proxy f -> Weigh ()

weighTriangle and weighLine for powers of 2 in experimentally derived ranges.

type Est = Estimate ConfInt Double

reportEstimate :: Report -> Est

estimate :: Benchmarkable -> IO Est

quickEstimate :: Benchmarkable -> IO Est

quickConfig :: Config

benchPow2s :: SemigroupFree f => String -> Proxy f -> Benchmark

benchTriangle and benchLine for powers of 2 in experimentally derived ranges.

benchable :: Applicative f => (forall g. SemigroupFree g => String -> Proxy g -> f a) -> f [a]

sequenceA a benchmark over all of the SemigroupFree implementations

weighFS :: Weigh ()

runWeighFS :: IO ()

benchCompare :: Int -> (forall g. SemigroupFree g => Proxy g -> Benchmarkable) -> IO ()

benchFS :: [Benchmark]

runBenchFS :: IO ()