Data.Tree.Utils

Documentation

countLeaves :: Tree a -> Int

Count the number of leaves of a tree

fromPath :: a -> [a] -> Tree a

Convert a non-empty path to a Tree

fromPathF :: [a] -> Forest a

Convert a possibly-empty path to a Forest

triangular :: Int -> Int

The n-th triangular number: sum [1..n]

triangularR :: Int -> Int -> Int

incParts :: Int -> Int -> Int -> Forest Int

The forest of increasing first-occurrence partitions (beginning with 1) with the given length incParts numberOfDifferences seriesLength seriesSum

boundarySeries :: Int -> Int -> Int -> Int -> Forest Int

The boundary series:

z,z..z, x..y, z,z..z Length is n

i before remainder after

 [X,z,z,z,z,z,z]
 [z,X,z,z,z,z,z]
 [z,z,X,z,z,z,z]
 [z,z,z,X,z,z,z]

trimTree :: Tree (Maybe a) -> Maybe (Tree a)

trimForest :: Forest (Maybe a) -> Forest a

Trim a forest. See trimTree

cycles :: Int -> Int -> Forest Int

cycles n m = All irreducible cycles of n elements of length m

A cycle cy is reducible iff there exists cy', 1 < n such that:

 replicate n cy' == cy

distinct0s :: Int -> Int -> Forest Int

distinct0s seems to have predictable polynomial lengths:

Are these just (length choose number_of_0s)?

λ> (n->mapM_ print$[getSum.foldMap(Sum. length.paths).distinct0s i$[i+1..i+n]|i<-[1..n]]) 12 [1,1,1,1,1,1,1,1,1,1,1,1] [1,2,3,4,5,6,7,8,9,10,11,12] [1,3,6,10,15,21,28,36,45,55,66,78] [1,4,10,20,35,56,84,120,165,220,286,364] [1,5,15,35,70,126,210,330,495,715,1001,1365] [1,6,21,56,126,252,462,792,1287,2002,3003,4368] [1,7,28,84,210,462,924,1716,3003,5005,8008,12376] [1,8,36,120,330,792,1716,3432,6435,11440,19448,31824] [1,9,45,165,495,1287,3003,6435,12870,24310,43758,75582] [1,10,55,220,715,2002,5005,11440,24310,48620,92378,167960] [1,11,66,286,1001,3003,8008,19448,43758,92378,184756,352716] [1,12,78,364,1365,4368,12376,31824,75582,167960,352716,705432] (41.30 secs, 21,512,577,928 bytes)

((n ==) . count (0 ==)) filter distincts 2 m == distinct0s n m distinct0s numberOfZeroes seriesLength

distincts :: Int -> Int -> Forest Int

distincts n m = the Tree of series of length m with n distinct elements ([0..n-1])

series :: Int -> Int -> Forest Int

series n m = the Tree of series of length m with up to n distinct elements ([0..n-1])

paths :: Tree a -> [[a]]

All paths in a Tree

stepFilter :: (a -> b) -> (a -> b -> b) -> (b -> Bool) -> Tree a -> Maybe (Tree a)

Given a way to convert an initial value to a state, a way to update the state with the new value, and a way to extract a result of whether to keep or not to keep the resulting series, filter a Tree

stepFilter1 :: (a -> b -> b) -> (b -> Bool) -> b -> Tree a -> Maybe (Tree a)

Step-fiter a Tree, given an initial state

stepFilterWith :: (a -> b -> b) -> (b -> Bool) -> b -> a -> Forest a -> Maybe (Tree a)

Step-fiter a piece-wise Tree, given an initial state

stepFilters :: (a -> b) -> (a -> b -> b) -> (b -> Bool) -> Forest a -> Forest a

Step-fiter a Forest, given an initial state function

stepFilters1 :: (a -> b -> b) -> (b -> Bool) -> b -> Forest a -> Forest a

Step-fiter a Forest, given an initial state

mapsTree :: (Forest a -> Forest a) -> Tree a -> Tree a

Map over all sub-Forests in a Tree

firstForest :: Tree a -> Tree (NonEmpty a)

Replace each node in a tree with the the NonEmpty list consisting of the original node, followed by the first path after that node.

When using drawTree or drawTreeU, this can be used to compress the representation, especially when there is little branching.

fseries :: Enum a => a -> Forest a

A Forest of all series with elements in 1..input

stepTree :: (a -> b -> b) -> (a -> b) -> Tree a -> Tree b

Scan along a Tree with a stepper and converter

stepTree1 :: (a -> b -> b) -> (a -> b) -> b -> Tree a -> Tree b

Scan along a Tree with a stepper, converter, and initial scan state

stepForest :: (a -> b -> b) -> (a -> b) -> b -> Forest a -> Forest b

Scan along a Forest with a stepper, converter, and initial scan state

stepTreeList :: (a -> b -> b) -> (a -> b) -> Tree a -> Tree ([a], b)

stepTree, returning a list of successive states

stepTreeList1 :: (a -> b -> b) -> (a -> b) -> b -> Tree a -> Tree ([a], b)

Scan along a Tree with a stepper, converter, and initial scan state

stepForestList :: (a -> b -> b) -> (a -> b) -> b -> Forest a -> Forest ([a], b)

Scan along a Forest with a stepper, converter, and initial scan state

unzipForest :: Forest a -> ([a], [Forest a])

Equivalent to:

 unzip . map ((Node x xs) -> (x, xs))

takeDepth :: Int -> Tree a -> Tree a

Take up to the given depth

takeDepths :: Int -> Forest a -> Forest a

Take up to the given depth

maxTreeOn_ :: Ord b => (a -> b) -> Tree a -> [a]

Maximum of a tree, using a comparison function

maxForestOn :: Ord b => (a -> b) -> Forest a -> Tree (b, a)

The maximum of a forest on a given value, returning that value.

This implementation assumes that computing the given function is expensive and only computes it once.

maxForestOn_ :: Ord b => (a -> b) -> Forest a -> Tree a

The maximum of a forest on a given value.

This implementation assumes that computing the given function is cheap.

drawTreeU :: Tree String -> String

From: https://github.com/haskell/containers/pull/344/files

Neat 2-dimensional drawing of a tree. Unicode paths

drawTreeU' :: Tree String -> String

More compact drawTreeU

drawForestU :: Forest String -> String

Neat 2-dimensional drawing of a forest. Unicode paths

drawU :: Tree String -> [String]

Draw a tree with spaces

drawU' :: Tree String -> [String]

Draw a tree with unicode

Orphan instances

Ord a => Ord (Tree a)

Instance details

Methods

compare :: Tree a -> Tree a -> Ordering

(<) :: Tree a -> Tree a -> Bool

(<=) :: Tree a -> Tree a -> Bool

(>) :: Tree a -> Tree a -> Bool

(>=) :: Tree a -> Tree a -> Bool

max :: Tree a -> Tree a -> Tree a

min :: Tree a -> Tree a -> Tree a